OCT-111: moved UpdateCatalogProductSelectionCriteriaQuery
