+++ components/catalogs/back/tests/Unit/Application/Handler/CreateCatalogHandlerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\Catalogs\Test\Unit\Application\Handler;
 
 use Akeneo\Catalogs\Application\Handler\CreateCatalogHandler;
-use Akeneo\Catalogs\Application\Persistence\UpdateCatalogProductSelectionCriteriaQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Catalog\UpdateCatalogProductSelectionCriteriaQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\UpsertCatalogQueryInterface;
 use Akeneo\Catalogs\ServiceAPI\Command\CreateCatalogCommand;
 use PHPUnit\Framework\TestCase;
